Run, report, and translate with pc_align for multi-processed altimetry
