+++ asp_plot/altimetry.py
@@ -10,17 +10,11 @@ import numpy as np
 import pandas as pd
 import rioxarray
 import xarray as xr
-from osgeo import gdal, osr
 from sliderule import icesat2
 
+from asp_plot.alignment import Alignment
 from asp_plot.stereopair_metadata_parser import StereopairMetadataParser
-from asp_plot.utils import (
-    ColorBar,
-    Raster,
-    glob_file,
-    run_subprocess_command,
-    save_figure,
-)
+from asp_plot.utils import ColorBar, Raster, glob_file, save_figure
 
 icesat2.init("slideruleearth.io", verbose=True)
 
@@ -274,15 +268,67 @@ class Altimetry:
 
             self.atl06sr_processing_levels_filtered[key] = atl06_filtered
 
-    def to_csv_for_pc_align(self, filename="atl06sr_for_pc_align"):
-        for key, atl06sr in self.atl06sr_processing_levels_filtered.items():
-            fn_out = f"{filename}_{key}.csv"
-            df = atl06sr[["geometry", "h_mean"]].copy()
-            df["lon"] = df["geometry"].x
-            df["lat"] = df["geometry"].y
-            df["height_above_datum"] = df["h_mean"]
-            df = df[["lon", "lat", "height_above_datum"]]
-            df.to_csv(fn_out, header=True, index=False)
+    def to_csv_for_pc_align(
+        self, key="high_confidence", filename_prefix="atl06sr_for_pc_align"
+    ):
+        atl06sr = self.atl06sr_processing_levels_filtered[key]
+        csv_fn = f"{filename_prefix}_{key}.csv"
+        df = atl06sr[["geometry", "h_mean"]].copy()
+        df["lon"] = df["geometry"].x
+        df["lat"] = df["geometry"].y
+        df["height_above_datum"] = df["h_mean"]
+        df = df[["lon", "lat", "height_above_datum"]]
+        df.to_csv(csv_fn, header=True, index=False)
+        return csv_fn
+
+    def alignment_report(
+        self,
+        processing_level="high_confidence",
+        minimum_points=500,
+        write_out_aligned_dem=False,
+        key_for_aligned_dem="high_confidence",
+    ):
+        filtered_keys = [
+            key
+            for key in self.atl06sr_processing_levels_filtered.keys()
+            if key.startswith(processing_level)
+        ]
+
+        alignment = Alignment(self.directory, self.dem_fn)
+
+        for key in filtered_keys:
+            atl06sr = self.atl06sr_processing_levels_filtered[key]
+            if atl06sr.shape[0] < minimum_points:
+                print(
+                    f"\n{key} has {atl06sr.shape[0]} points, which is less than the suggested {minimum_points} points. Skipping alignment.\n"
+                )
+                continue
+
+            if not os.path.exists(
+                glob_file(
+                    os.path.join(self.directory, "pc_align"), f"*{key}*transform.txt"
+                )
+            ):
+                csv_fn = self.to_csv_for_pc_align(key=key)
+
+                print(f"\nAligning {key} to DEM with pc_align\n")
+
+                alignment.pc_align_dem_to_atl06sr(
+                    atl06sr_csv=csv_fn,
+                    output_prefix=f"pc_align/pc_align_{key}",
+                )
+
+        for key in filtered_keys:
+            report = alignment.pc_align_report(output_prefix=f"pc_align/pc_align_{key}")
+            print(f"\n{key} alignment report:\n{report}\n")
+
+        if write_out_aligned_dem:
+            self.aligned_dem_fn = alignment.apply_dem_translation(
+                output_prefix=f"pc_align/pc_align_{key_for_aligned_dem}",
+            )
+            print(
+                f"\nWrote out {key_for_aligned_dem} aligned DEM to {self.aligned_dem_fn}\n"
+            )
 
     def plot_atl06sr_time_stamps(
         self,
@@ -437,203 +483,6 @@ class Altimetry:
         if save_dir and fig_fn:
             save_figure(fig, save_dir, fig_fn)
 
-    # TODO: move all pc_align steps and dem translations to separate class
-    # call this alignment for different processing levels, with some minimum number
-    # of required points (a ~500 point parameter).
-    # Report all translation results to user via a dictionary and printing.
-    # If translation agrees within some tolerance (a 5% or 10% parameter) in XYZ components,
-    # apply the translation of the points found closest in time to the DEM.
-    def pc_align_dem_to_atl06sr(
-        self,
-        max_displacement=20,
-        max_source_points=10000000,
-        alignment_method="point-to-point",
-        atl06sr_csv=None,
-        output_prefix=None,
-    ):
-        if atl06sr_csv is None or not os.path.exists(atl06sr_csv):
-            raise ValueError(
-                f"\nATL06 filtered CSV file not found: {atl06sr_csv}\n\nWe need this to run pc_align.\n"
-            )
-        if not output_prefix:
-            raise ValueError("\nPlease provide an output prefix for pc_align.\n")
-        if self.aligned_dem_fn:
-            logger.warning(
-                f"\nAligned DEM was already supplied: {self.aligned_dem_fn}\n\nPlease use that, or remove this file before running pc_align.\n"
-            )
-            return
-
-        command = [
-            "pc_align",
-            "--max-displacement",
-            str(max_displacement),
-            "--max-num-source-points",
-            str(max_source_points),
-            "--alignment-method",
-            alignment_method,
-            "--csv-format",
-            "1:lon 2:lat 3:height_above_datum",
-            "--compute-translation-only",
-            "--output-prefix",
-            output_prefix,
-            self.dem_fn,
-            atl06sr_csv,
-        ]
-
-        run_subprocess_command(command)
-
-    def pc_align_report(self, pc_align_folder="pc_align"):
-        pc_align_log = glob_file(pc_align_folder, "*log-pc_align*.txt")
-
-        with open(pc_align_log, "r") as file:
-            content = file.readlines()
-
-        report = ""
-        for line in content:
-            if "Input: error percentile of smallest errors (meters):" in line:
-                report_line = line.split("[ console ] : ")[1]
-                report += report_line
-            if "Input: mean of smallest errors (meters):" in line:
-                report_line = line.split("[ console ] : ")[1]
-                report += report_line
-            if "Output: error percentile of smallest errors (meters):" in line:
-                report_line = line.split("[ console ] : ")[1]
-                report += report_line
-            if "Output: mean of smallest errors (meters):" in line:
-                report_line = line.split("[ console ] : ")[1]
-                report += report_line
-            if "Translation vector (Cartesian, meters):" in line:
-                report_line = line.split("[ console ] : ")[1]
-                report += report_line
-            if "Translation vector magnitude (meters):" in line:
-                report_line = line.split("[ console ] : ")[1]
-                report += report_line
-
-        return report
-
-    def generate_translated_dem(self, pc_align_output, dem_out_fn):
-        if not os.path.exists(pc_align_output):
-            raise ValueError(
-                f"\npc_align output not found: {pc_align_output}\n\nWe need this to generate the translated DEM.\n"
-            )
-        if self.aligned_dem_fn:
-            logger.warning(
-                f"\nAligned DEM already exists: {self.aligned_dem_fn}\n\nPlease use that, or remove this file before running pc_align.\n"
-            )
-            return
-
-        rast = Raster(self.dem_fn)
-        gsd = rast.get_gsd()
-        epsg = rast.get_epsg_code()
-
-        command = [
-            "point2dem",
-            "--tr",
-            str(gsd),
-            "--t_srs",
-            f"EPSG:{epsg}",
-            "--nodata-value",
-            str(-9999),
-            "-o",
-            dem_out_fn,
-            pc_align_output,
-        ]
-
-        run_subprocess_command(command)
-
-    def apply_dem_translation(self, pc_align_folder="pc_align", inv_trans=True):
-        def get_proj_shift(src_c, src_shift, s_srs, t_srs, inv_trans=True):
-            if s_srs.IsSame(t_srs):
-                proj_shift = src_shift
-            else:
-                src_c_shift = src_c + src_shift
-                src2proj = osr.CoordinateTransformation(s_srs, t_srs)
-                proj_c = np.array(src2proj.TransformPoint(*src_c))
-                proj_c_shift = np.array(src2proj.TransformPoint(*src_c_shift))
-                if inv_trans:
-                    proj_shift = proj_c - proj_c_shift
-                else:
-                    proj_shift = proj_c_shift - proj_c
-            # Reduce unnecessary precision
-            proj_shift = np.around(proj_shift, 3)
-            return proj_shift
-
-        if self.aligned_dem_fn:
-            logger.warning(
-                f"\nAligned DEM already exists: {self.aligned_dem_fn}\n\nPlease use that, or remove this file before running pc_align.\n"
-            )
-            return
-
-        pc_align_log = glob_file(pc_align_folder, "*log-pc_align*.txt")
-
-        src = Raster(self.dem_fn)
-        src_a = src.read_array()
-        src_ndv = src.get_ndv()
-
-        # Need to extract from log to know how to compute translation
-        # if ref is csv and src is dem, want to transform source_center + shift
-        # if ref is dem and src is csv, want to inverse transform ref by shift applied at (source_center - shift)
-
-        llz_c = None
-        with open(pc_align_log, "r") as file:
-            content = file.readlines()
-
-        for line in content:
-            if "Centroid of source points (Cartesian, meters):" in line:
-                ecef_c = np.genfromtxt([line.split("Vector3")[1][1:-2]], delimiter=",")
-            if "Centroid of source points (lat,lon,z):" in line:
-                llz_c = np.genfromtxt([line.split("Vector3")[1][1:-2]], delimiter=",")
-            if "Translation vector (Cartesian, meters):" in line:
-                ecef_shift = np.genfromtxt(
-                    [line.split("Vector3")[1][1:-2]], delimiter=","
-                )
-            if "Translation vector (lat,lon,z):" in line:
-                llz_shift = np.genfromtxt(
-                    [line.split("Vector3")[1][1:-2]], delimiter=","
-                )
-                break
-
-        if llz_c is None:
-            raise ValueError(
-                f"\nLog file does not contain necessary translation information: {pc_align_log}\n"
-            )
-
-        # Reorder lat,lon,z to lon,lat,z (x,y,z)
-        i = [1, 0, 2]
-        llz_c = llz_c[i]
-        llz_shift = llz_shift[i]
-
-        ecef_srs = osr.SpatialReference()
-        ecef_srs.ImportFromEPSG(4978)
-
-        s_srs = ecef_srs
-        src_c = ecef_c
-        src_shift = ecef_shift
-
-        # Determine shift in original dataset coords
-        t_srs = osr.SpatialReference()
-        t_srs.ImportFromWkt(src.ds.crs.to_wkt())
-        proj_shift = get_proj_shift(src_c, src_shift, s_srs, t_srs, inv_trans)
-
-        self.aligned_dem_fn = self.dem_fn.replace(".tif", "_pc_align_translated.tif")
-        gdal_opt = ["COMPRESS=LZW", "TILED=YES", "PREDICTOR=3", "BIGTIFF=IF_SAFER"]
-        dst_ds = gdal.GetDriverByName("GTiff").CreateCopy(
-            self.aligned_dem_fn, gdal.Open(self.dem_fn), strict=0, options=gdal_opt
-        )
-        # Apply vertical shift
-        dst_b = dst_ds.GetRasterBand(1)
-        dst_b.SetNoDataValue(float(src_ndv))
-        dst_b.WriteArray(np.around((src_a + proj_shift[2]).filled(src_ndv), decimals=3))
-
-        dst_gt = list(dst_ds.GetGeoTransform())
-        # Apply horizontal shift directly to geotransform
-        dst_gt[0] += proj_shift[0]
-        dst_gt[3] += proj_shift[1]
-        dst_ds.SetGeoTransform(dst_gt)
-
-        print(f"\nWriting out: {self.aligned_dem_fn}\n")
-        dst_ds = None
-
     def atl06sr_to_dem_dh(self):
         dem = rioxarray.open_rasterio(self.dem_fn, masked=True).squeeze()
         epsg = dem.rio.crs.to_epsg()
@@ -677,6 +526,9 @@ class Altimetry:
     ):
         if plot_aligned:
             column_name = "icesat_minus_aligned_dem"
+            if not self.aligned_dem_fn:
+                print("\nAligned DEM not found.\n")
+                return
         else:
             column_name = "icesat_minus_dem"
 
@@ -732,6 +584,9 @@ class Altimetry:
         column_names = ["icesat_minus_dem"]
         if plot_aligned:
             column_names.append("icesat_minus_aligned_dem")
+            if not self.aligned_dem_fn:
+                print("\nAligned DEM not found.\n")
+                return
 
         fig, ax = plt.subplots(1, 1, figsize=(6, 4))
